#47: Changed AcceptOnceUnchangedFileListFilter into a simpler AcceptUnchangedFileListFilter
