+++ .gitignore
@@ -15,3 +15,4 @@ target/*
 
 *.iml
 .idea/*
+.run/*
