+++ src/test/java/fi/fmi/avi/archiver/spring/integration/file/filters/AcceptUnchangedFileListFilterTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 import org.springframework.integration.file.filters.CompositeFileListFilter;
 import org.springframework.integration.file.filters.ReversibleFileListFilter;
-import org.springframework.test.util.ReflectionTestUtils;
 
 import java.io.File;
 import java.io.IOException;
@@ -17,29 +16,28 @@ import java.nio.file.attribute.FileTime;
 import java.time.Instant;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
-import java.util.Queue;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 
-public class AcceptOnceUnchangedFileListFilterTest {
+public class AcceptUnchangedFileListFilterTest {
 
     @Test
     public void test(@TempDir Path tempDir) throws IOException {
-        final AcceptOnceUnchangedFileListFilter filter = new AcceptOnceUnchangedFileListFilter();
+        final AcceptUnchangedFileListFilter filter = new AcceptUnchangedFileListFilter();
         final Path tempFile = tempDir.resolve("temp");
         final File file = tempFile.toFile();
 
         assertThat(filter.accept(file)).isFalse();
         assertThat(filter.accept(file)).isTrue();
+        // File is removed from seen map
         assertThat(filter.accept(file)).isFalse();
-        assertThat(filter.accept(file)).isFalse();
+        assertThat(filter.accept(file)).isTrue();
     }
 
     @Test
     public void test_content_change(@TempDir Path tempDir) throws IOException {
-        final AcceptOnceUnchangedFileListFilter filter = new AcceptOnceUnchangedFileListFilter();
+        final AcceptUnchangedFileListFilter filter = new AcceptUnchangedFileListFilter();
         final Path tempFile = tempDir.resolve("temp");
         Files.write(tempFile, "temp content".getBytes(StandardCharsets.UTF_8));
         final File file = tempFile.toFile();
@@ -48,12 +46,11 @@ public class AcceptOnceUnchangedFileListFilterTest {
         Files.write(tempFile, "more content".getBytes(StandardCharsets.UTF_8), StandardOpenOption.APPEND);
         assertThat(filter.accept(file)).isFalse();
         assertThat(filter.accept(file)).isTrue();
-        assertThat(filter.accept(file)).isFalse();
     }
 
     @Test
     public void test_multiple_content_changes(@TempDir Path tempDir) throws IOException {
-        final AcceptOnceUnchangedFileListFilter filter = new AcceptOnceUnchangedFileListFilter();
+        final AcceptUnchangedFileListFilter filter = new AcceptUnchangedFileListFilter();
         final Path tempFile = tempDir.resolve("temp");
         Files.write(tempFile, "temp content".getBytes(StandardCharsets.UTF_8));
         final File file = tempFile.toFile();
@@ -70,7 +67,7 @@ public class AcceptOnceUnchangedFileListFilterTest {
 
     @Test
     public void test_last_modified_change(@TempDir Path tempDir) throws IOException {
-        final AcceptOnceUnchangedFileListFilter filter = new AcceptOnceUnchangedFileListFilter();
+        final AcceptUnchangedFileListFilter filter = new AcceptUnchangedFileListFilter();
         final Path tempFile = tempDir.resolve("temp");
         Files.write(tempFile, "temp content".getBytes(StandardCharsets.UTF_8));
         Files.setLastModifiedTime(tempFile, FileTime.from(Instant.parse("2020-05-18T12:00:00Z")));
@@ -85,7 +82,7 @@ public class AcceptOnceUnchangedFileListFilterTest {
 
     @Test
     public void test_multiple_last_modified_changes(@TempDir Path tempDir) throws IOException {
-        final AcceptOnceUnchangedFileListFilter filter = new AcceptOnceUnchangedFileListFilter();
+        final AcceptUnchangedFileListFilter filter = new AcceptUnchangedFileListFilter();
         final Path tempFile = tempDir.resolve("temp");
         Files.write(tempFile, "temp content".getBytes(StandardCharsets.UTF_8));
         Files.setLastModifiedTime(tempFile, FileTime.from(Instant.parse("2020-05-18T12:00:00Z")));
@@ -103,7 +100,7 @@ public class AcceptOnceUnchangedFileListFilterTest {
 
     @Test
     public void test_filter_files(@TempDir Path tempDir) throws IOException {
-        final AcceptOnceUnchangedFileListFilter filter = new AcceptOnceUnchangedFileListFilter();
+        final AcceptUnchangedFileListFilter filter = new AcceptUnchangedFileListFilter();
         final Path tempFile = tempDir.resolve("temp");
         Files.write(tempFile, "temp content".getBytes(StandardCharsets.UTF_8));
         final File file = tempFile.toFile();
@@ -111,13 +108,14 @@ public class AcceptOnceUnchangedFileListFilterTest {
         assertThat(filter.filterFiles(new File[]{file})).hasSize(0);
         assertThat(filter.filterFiles(new File[]{file})).hasSize(1);
         assertThat(filter.filterFiles(new File[]{file})).hasSize(0);
-        assertThat(filter.filterFiles(new File[]{file})).hasSize(0);
+        assertThat(filter.filterFiles(new File[]{file})).hasSize(1);
         assertThat(filter.accept(file)).isFalse();
+        assertThat(filter.accept(file)).isTrue();
     }
 
     @Test
     public void test_filter_files_modification(@TempDir Path tempDir) throws IOException {
-        final AcceptOnceUnchangedFileListFilter filter = new AcceptOnceUnchangedFileListFilter();
+        final AcceptUnchangedFileListFilter filter = new AcceptUnchangedFileListFilter();
         final Path tempFile = tempDir.resolve("temp");
         Files.write(tempFile, "temp content".getBytes(StandardCharsets.UTF_8));
         final File file = tempFile.toFile();
@@ -127,60 +125,29 @@ public class AcceptOnceUnchangedFileListFilterTest {
         assertThat(filter.filterFiles(new File[]{file})).hasSize(0);
         assertThat(filter.filterFiles(new File[]{file})).hasSize(1);
         assertThat(filter.filterFiles(new File[]{file})).hasSize(0);
-        assertThat(filter.filterFiles(new File[]{file})).hasSize(0);
+        assertThat(filter.filterFiles(new File[]{file})).hasSize(1);
         assertThat(filter.accept(file)).isFalse();
-    }
-
-    @Test
-    @SuppressWarnings("unchecked")
-    public void test_capacity(@TempDir Path tempDir) {
-        final AcceptOnceUnchangedFileListFilter filter = new AcceptOnceUnchangedFileListFilter(2);
-        final File fooFile = tempDir.resolve("foo").toFile();
-        final File barFile = tempDir.resolve("bar").toFile();
-        final File bazFile = tempDir.resolve("baz").toFile();
-
-        assertThat(filter.accept(fooFile)).isFalse();
-        assertThat(filter.accept(fooFile)).isTrue();
-        assertThat(filter.accept(barFile)).isFalse();
-        assertThat(filter.accept(barFile)).isTrue();
-        assertThat(filter.accept(fooFile)).isFalse();
-
-        final Queue<File> seen = (Queue<File>) ReflectionTestUtils.getField(filter, "seen");
-        assertThat(seen).hasSize(2);
-
-        // Baz overflows the queue
-        assertThat(filter.accept(bazFile)).isFalse();
-        assertThat(filter.accept(bazFile)).isTrue();
-
-        // Foo can be accepted again
-        assertThat(filter.accept(fooFile)).isFalse();
-        assertThat(filter.accept(fooFile)).isTrue();
-
-        final Map<File, AcceptOnceUnchangedFileListFilter.FileData> seenMap =
-                (Map<File, AcceptOnceUnchangedFileListFilter.FileData>) ReflectionTestUtils.getField(filter, "seenMap");
-        assertThat(seenMap).isNotNull();
-        assertThat(seenMap).hasSize(2);
-        assertThat(seen).containsExactly(bazFile, fooFile);
-        assertThat(seenMap.keySet()).contains(fooFile, bazFile);
+        assertThat(filter.accept(file)).isTrue();
     }
 
     @Test
     public void test_rollback(@TempDir Path tempDir) {
-        final AcceptOnceUnchangedFileListFilter filter = new AcceptOnceUnchangedFileListFilter();
+        final AcceptUnchangedFileListFilter filter = new AcceptUnchangedFileListFilter();
         testRollback(filter, tempDir);
     }
 
     @Test
     public void test_rollback_composite(@TempDir Path tempDir) {
-        final AcceptOnceUnchangedFileListFilter filter = new AcceptOnceUnchangedFileListFilter();
+        final AcceptUnchangedFileListFilter filter = new AcceptUnchangedFileListFilter();
         final CompositeFileListFilter<File> composite = new CompositeFileListFilter<>(Collections.singletonList(filter));
         testRollback(composite, tempDir);
     }
 
     private void testRollback(final ReversibleFileListFilter<File> filter, final Path tempDir) {
+        final File foo = tempDir.resolve("foo").toFile();
         final File bar = tempDir.resolve("bar").toFile();
         final File baz = tempDir.resolve("baz").toFile();
-        final File[] files = new File[]{tempDir.resolve("foo").toFile(), bar, baz};
+        final File[] files = new File[]{foo, bar, baz};
 
         List<File> passed = filter.filterFiles(files);
         assertThat(passed).isEmpty();
@@ -191,16 +158,13 @@ public class AcceptOnceUnchangedFileListFilterTest {
         assertThat(passedNow).hasSize(0);
         filter.rollback(passed.get(1), passed);
         passedNow = filter.filterFiles(files);
-        assertThat(passedNow).hasSize(0);
+        assertThat(passedNow).hasSize(1);
+        assertThat(passedNow.get(0)).isEqualTo(foo);
         passedNow = filter.filterFiles(files);
         assertThat(passedNow).hasSize(2);
 
         assertThat(passedNow.get(0)).isEqualTo(bar);
         assertThat(passedNow.get(1)).isEqualTo(baz);
-        passedNow = filter.filterFiles(files);
-        assertThat(passedNow).hasSize(0);
-        passedNow = filter.filterFiles(files);
-        assertThat(passedNow).hasSize(0);
     }
 
 }
